+++ HISTORY.rst
@@ -3,6 +3,18 @@
 Release History
 ---------------
 
+2.5.5 (2023-12-11)
++++++++++++++++++++
+
+**Improvements**
+
+- #725 Remove redundant event logging
+- Refactor _process_sports_data to use eventId
+
+**Bug Fixes**
+
+- Correctly handle LINE_RANGE markets within trading controls / blotter / profit / order
+
 2.5.4 (2023-11-13)
 +++++++++++++++++++